linux: remove device tree patch file and linux 6.6 recipe extension
Now that we have a way to install device tree source files into the right
location within the linux tree, we can also install the main sa67 device
tree source file the same way. This will allow as to get rid of the
patch file. Add the device tree source file and remove references to the
patch file along with the patch file itself and the linux 6.6 recipe
appendix which still relies on the patch.

Signed-off-by: Thomas Schaefer <thomas.schaefer@kontron.com>