uboot: add EXTLINUX support for sa67-mainline machine
Define the necessary EXTLINUX configurations to create an appropriate
extlinux.conf file that is stored to the boot partition. Make sure that
the boot files (kernel image, devicetree and devicetree overlays) are
copied to the boot partition as well by adding them to the
IMAGE_BOOT_FILES list. Inherit the uboot-extlinux-config in the u-boot
mainline recipe.

Note that the current extlinux class does not support applying device
tree overlays when booting.

Signed-off-by: Thomas Schaefer <thomas.schaefer@kontron.com>